refactoring files project files
