+++ src/auth/base_config.py
@@ -1,15 +1,13 @@
-import os
-
+from fastapi_users import FastAPIUsers
 from fastapi_users.authentication import CookieTransport, AuthenticationBackend
 from fastapi_users.authentication import JWTStrategy
 
-from dotenv import load_dotenv
-load_dotenv(override=True)
+from src.auth.models import User
+from src.auth.manager import get_user_manager
+from src.config import SECRET
+
 
 cookie_transport = CookieTransport(cookie_name='s', cookie_max_age=3600)
-
-
-SECRET = os.getenv("SECRET")
 
 
 def get_jwt_strategy() -> JWTStrategy:
@@ -21,3 +19,10 @@ auth_backend = AuthenticationBackend(
     transport=cookie_transport,
     get_strategy=get_jwt_strategy,
 )
+
+fastapi_users = FastAPIUsers[User, int](
+    get_user_manager,
+    [auth_backend],
+)
+
+current_user = fastapi_users.current_user()
