+++ src/auth/manager.py
@@ -1,15 +1,12 @@
-import os
 from typing import Optional
 
 from fastapi import Depends, Request
 from fastapi_users import BaseUserManager, IntegerIDMixin
 from fastapi_users import exceptions, models, schemas
-from auth.database import User, get_user_db
 
-from dotenv import load_dotenv
-load_dotenv(override=True)
-
-SECRET = os.getenv("SECRET")
+from src.auth.models import User
+from src.auth.utils import get_user_db
+from src.config import SECRET
 
 
 class UserManager(IntegerIDMixin, BaseUserManager[User, int]):
