+++ src/auth/models.py
@@ -1,10 +1,19 @@
 from datetime import datetime
+from fastapi_users.db import SQLAlchemyBaseUserTable
+
+from sqlalchemy.orm import Mapped, mapped_column
 
 from sqlalchemy import (
     MetaData, Table, Column, Integer, String, TIMESTAMP, ForeignKey, JSON, Boolean
 )
+from sqlalchemy.orm import DeclarativeBase
+# from src.database import Base
 
 metadata = MetaData()
+
+
+class Base(DeclarativeBase):
+    pass
 
 # role = Table(
 #     "role",
@@ -13,6 +22,7 @@ metadata = MetaData()
 #     Column("name", String, nullable=False),
 #     Column("permissions", JSON),
 # )
+
 
 user = Table(
     "user",
@@ -27,3 +37,17 @@ user = Table(
     Column("is_superuser", Boolean, default=False, nullable=False),
     Column("is_verified", Boolean, default=False, nullable=False),
 )
+
+
+class User(SQLAlchemyBaseUserTable[int], Base):
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+
+    username: Mapped[str] = mapped_column(String, nullable=False)
+    registered_at: Mapped[str] = mapped_column(TIMESTAMP, default=datetime.utcnow)
+
+    email: Mapped[str] = mapped_column(String(length=320), unique=True, index=True, nullable=False)
+    hashed_password: Mapped[str] = mapped_column(String(length=1024), nullable=False)
+    is_active: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False)
+    is_superuser: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
+    is_verified: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
+
