+++ src/config.py
@@ -4,6 +4,8 @@ from dotenv import load_dotenv
 
 load_dotenv(override=True)
 
+SECRET = os.getenv("SECRET")
+
 DB_NAME = os.getenv("DB_NAME")
 DB_USER = os.getenv("DB_USER")
 DB_PASS = os.getenv("DB_PASS")
