+++ src/database.py
@@ -1,55 +1,16 @@
-from datetime import datetime
 from typing import AsyncGenerator
 
-from fastapi import Depends
-from fastapi_users.db import SQLAlchemyBaseUserTable, SQLAlchemyBaseUserTableUUID, SQLAlchemyUserDatabase
-from sqlalchemy import TIMESTAMP, String, Boolean, Integer
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
 from config import DB_HOST, DB_NAME, DB_PASS, DB_PORT, DB_USER
 
 # DATABASE_URL = "sqlite+aiosqlite:///./test.db"
 DATABASE_URL = f"postgresql+asyncpg://{DB_USER}:{DB_PASS}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
 
-
-class Base(DeclarativeBase):
-    pass
-
-
-# class User(SQLAlchemyBaseUserTableUUID, Base):
-#     pass
-
-# By default, we use UUID as a primary key ID for your user.
-# If you want to use another type, like an auto-incremented integer,
-# you can use SQLAlchemyBaseUserTable as base class and define your own id column.
-# Notice that SQLAlchemyBaseUserTable expects a generic type to define the actual type of ID you use.
-class User(SQLAlchemyBaseUserTable[int], Base):
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-
-    username: Mapped[str] = mapped_column(String, nullable=False)
-    registered_at: Mapped[str] = mapped_column(TIMESTAMP, default=datetime.utcnow)
-
-    email: Mapped[str] = mapped_column(String(length=320), unique=True, index=True, nullable=False)
-    hashed_password: Mapped[str] = mapped_column(String(length=1024), nullable=False)
-    is_active: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False)
-    is_superuser: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
-    is_verified: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
-
-
 engine = create_async_engine(DATABASE_URL)
 async_session_maker = async_sessionmaker(engine, expire_on_commit=False)
-
-
-# async def create_db_and_tables():
-#     async with engine.begin() as conn:
-#         await conn.run_sync(Base.metadata.create_all)
 
 
 async def get_async_session() -> AsyncGenerator[AsyncSession, None]:
     async with async_session_maker() as session:
         yield session
-
-
-async def get_user_db(session: AsyncSession = Depends(get_async_session)):
-    yield SQLAlchemyUserDatabase(session, User)
