+++ src/main.py
@@ -1,19 +1,12 @@
 import uvicorn
 from fastapi import FastAPI, Depends
-from fastapi_users import FastAPIUsers
-# from fastapi_users import fastapi_users
 
-from auth.auth import auth_backend
-from auth.manager import get_user_manager
+from auth.base_config import auth_backend, fastapi_users
 from auth.schemas import UserRead, UserCreate
-from auth.database import User
+
 
 app = FastAPI()
 
-fastapi_users = FastAPIUsers[User, int](
-    get_user_manager,
-    [auth_backend],
-)
 
 app.include_router(
     fastapi_users.get_auth_router(auth_backend), prefix="/auth/jwt", tags=["auth"]
@@ -24,42 +17,12 @@ app.include_router(
     prefix="/auth",
     tags=["auth"],
 )
-# app.include_router(
-#     fastapi_users.get_reset_password_router(),
-#     prefix="/auth",
-#     tags=["auth"],
-# )
-# app.include_router(
-#     fastapi_users.get_verify_router(UserRead),
-#     prefix="/auth",
-#     tags=["auth"],
-# )
-# app.include_router(
-#     fastapi_users.get_users_router(UserRead, UserUpdate),
-#     prefix="/users",
-#     tags=["users"],
-# )
-#
-#
-# @app.get("/authenticated-route")
-# async def authenticated_route(user: User = Depends(current_active_user)):
-#     return {"message": f"Hello {user.email}!"}
 
 
-# ====== Данное поле нужно если не используем библиотеку Alembic ======
-# @app.on_event("startup")
-# async def on_startup():
-#     # Not needed if you setup a migration system like Alembic
-#     # Не требуется, если вы настроили систему миграции, например Alembic.
-#     await create_db_and_tables()
-
-current_user = fastapi_users.current_user()
-
-
-@app.get("/protected-route")
-def protected_route(user: User = Depends(current_user)):
-    return f"Hello, {user.email}"
+# @app.get("/protected-route")
+# def protected_route(user: User = Depends(current_user)):
+#     return f"Hello, {user.email}"
 
 
 if __name__ == "__main__":
-    uvicorn.run("main:app", host="0.0.0.0", log_level="info", reload=True)
+    uvicorn.run("main:app", log_level="info", reload=True)
